py-hprof: initial commit
The intention is to make a nice little library to enable scripting and
interactive exploration of .hprof files (primarily from Android, but
Java should be very similar). We'll be handling heap dumps; anything
else is out of scope.

Does basically nothing yet. There is just some scaffolding for binary
random access and streaming, reading ascii text, ints and uints, and
some tests for the above.