reduce risk of name collisions when accessing object fields
First of all, this commit introduces a test that checks all Allocation
subclasses for names not beginning with hprof, _hprof, or __.

Then, to make that test pass, we change lots and lots of names.

Finally, we remove some fields (like 'tag') that were not really
needed, and make others (like 'datalen') internal.