remove hprof.record.Unhandled, count unhandled tags instead
I thought it might be nice to expose the data of unhandled records
for extensibility, but realized that we could do that in a nicer way
by allowing users to register a custom parser for that record type.

Let's add that part when the need arises. Optimally, we'd want such
improvements to be contributed into the package anyway.