add support for parsing heap items, but don't save anything
This is the first step towards actually doing something useful.

The acceptance tests jumped from 3 seconds (of which 2 were bz2
decompression) to around 60. That's unreasonable. Let's see if
it can be optimized; otherwise, we'll have to switch to some
form of on-demand item parsing.

Soundtrack: The Hellacopters - Leave it alone