add demo/experiment of how to handle overloaded java attrs
...in the form of unit tests that actually test python behavior,
rather than the library's. Oh well.

This is important for the way I want to implement overloaded Java
attributes -- i.e. when a class and subclass declare the same attr.
I was worried that I might have to do something really complicated
to separate the namespaces, but it seems I'll be able to get away
with wrappers indicating which type of reference we have.

Soundtrack: Kyle Gass Band - Mama's Ma