add JavaClassName class, make class access more indirect
The previous model, where we had a tree of packages and classes, could
not handle duplicate classes (e.g. in different class loaders) in a
good and consistent way.

Instead, build a tree of JavaPackage and JavaClassName objects. The
class names will be used to look up class definitions; at that point,
there will be a list, and the caller can decide what to do when there
are multiple matches.

I would really prefer the old way; it would make for much cleaner code
when using the library. Maybe we can add a convenience function later.

Soundtrack: Jethro Tull - Locomotive Breath