resolve heap references
This means that we should now be able to follow all kinds of refs;
from object arrays, instance attributes, and static attributes.

It also means that you're guaranteed to get an error while parsing if
an object is missing. Perhaps a little slower than looking things up
as they're needed, but it's nice to know that the file you're looking
at is complete.

Soundtrack: Ozzy Osbourne - It's A Raid