defer instantiations 1/3: primitive arrays
Turns out I can't expect class dumps to come before instance dumps that
need those classes. This means we must defer instantiations until we
are sure that the classes are available.

This introduces some overhead (having to create tuples and add them to
the deferred list), but at least for tuples, it doesn't seem too bad.
It's possible to _try_ immediate instantiation, and only defer when
necessary, but I chose to keep the code simpler for now -- we can
change it later, if it turns out to make a significant difference.

Soundtrack: Robyn - Cobrastyle