add Heap.exact_instances() and Heap.all_instances()
These functions return iterables that yield all instances of a class,
excluding/including subclasses, respectively.

The data is kept in Heap._instances, which must be populated when
constructing the objects.

As usual, there is special handling for java.lang.Class; every class
is a technically a Class object, but doesn't really show up as such
in the dump format. The hierarchy is a bit tangled at the top there.

Soundtrack: The Darkness - Get Your Hands off My Woman